SkewLinearAlgebra.jl Documentation

SkewLinearAlgebra.skewhermitian — Method

skewhermitian(A)

Returns the skew-Hermitian part of A, i.e. (A-A')/2. See also skewhermitian!, which does this in-place.

SkewLinearAlgebra.skewhermitian! — Method

skewhermitian(A)

Transforms A in-place to its skew-Hermitian part (A-A')/2, and returns a SkewHermitian view.

SkewLinearAlgebra.isskewhermitian — Method

isskewhermitian(A)

Returns whether A is skew-Hermitian, i.e. whether A == -A'.

SkewLinearAlgebra.SkewHermitian — Method

SkewHermitian(A) <: AbstractMatrix

Construct a SkewHermitian view of the skew-Hermitian matrix A (A == -A'), which allows one to exploit efficient operations for eigenvalues, exponentiation, and more.

Takes "ownership" of the matrix A. See also skewhermitian, which takes the skew-hermitian part of A, and skewhermitian!, which does this in-place, along with isskewhermitian which checks whether A == -A'.

SkewLinearAlgebra.skewchol — Method

skewchol(A)

<<<<<<< HEAD Computes a Cholesky-like factorization of A real skew-symmetric. The function returns a SkewCholesky structure composed of three arguments: Rm,Jm,Pv. Rm is UpperTriangular, Jm is SkewHermTridiagonal, Pv is an array of integers. Let S be the returned structure, then the factorization is such that:

transpose(S.Rm)*S.Jm*S.Rm = A[S.Pv,S.Pv]

This factorization is issued from P. Benner et al, 
"[Cholesky-like factorizations of skew-symmetric matrices](https://etna.ricam.oeaw.ac.at/vol.11.2000/pp85-93.dir/pp85-93.pdf)"(2000). 

=======

Computes a Cholesky-like factorization of the real skew-symmetric matrix A. The function returns a SkewCholesky structure composed of three fields: Rm,Jm,Pv. Rm is UpperTriangular, Jm is SkewHermTridiagonal, Pv is an array of integers. Let S be the returned structure, then the factorization is such that S.Rm'*S.Jm*S.Rm = A[S.Pv,S.Pv]

This factorization (and the underlying algorithm) is described in from P. Benner et al, "Cholesky-like factorizations of skew-symmetric matrices"(2000).

338103351bd235f2d1b76117f98920182551c96b